+++ tools/uninstall_pod.sh
@@ -4,6 +4,7 @@
 
 NAME=${NAME-"reds"}
 PREFIX=${PREFIX-"/usr/local"}
+RMLIB=${RMLIB-"auto"}
 
 BINPATH=${BINPATH-"${PREFIX}/bin"}
 LIBPATH=${LIBPATH-"${PREFIX}/lib"}
@@ -24,10 +25,12 @@ sudo -u postgres psql -c "DROP USER \"${POSTGRESQL_ROLE}\";"
 
 # INFO Remove files
 
-rm "${BINFILE}"
-rm "${ETCFILE}"
-rm "${LOGFILE}"
-if [ ! `ls -A "${ETCPATH}/reds"` ]; then
-    rm -r "${ETCPATH}/reds"
-    rm -r "${LIBPATH}/reds"
+[ -f "${BINFILE}" ] && rm "${BINFILE}"
+[ -f "${ETCFILE}" ] && rm "${ETCFILE}"
+[ -f "${LOGFILE}" ] && rm "${LOGFILE}"
+
+[ $RMLIB = auto ] && [ -d "${ETCPATH}/reds" ] && [ -z "`ls -A "${ETCPATH}/reds"`" ] && RMLIB=true
+if [ $RMLIB = true ]; then
+    [ -d "${ETCPATH}/reds" ] && rm -r "${ETCPATH}/reds"
+    [ -d "${LIBPATH}/reds" ] && rm -r "${LIBPATH}/reds"
 fi
